requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
GeoAlchemy2==0.13.1
greenlet==2.0.2
idna==3.4
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
wtforms-validators==1.0.0
